Fix up loads of types so that more advanced tests work :)

+++ src/stack.c
@@ -6,7 +6,7 @@
 
 
 
-int init_frame(frame_t *prev, int max_stack_size, unsigned short max_local_size, int pc, short n_args) {
+int init_frame(frame_t *prev, int max_stack_size, unsigned short max_local_size, int pc, unsigned short n_args) {
     frame_t *new_frame = (frame_t *)malloc(sizeof(frame_t));
 
     new_frame->local_data = (word_t *)malloc(max_local_size * sizeof(word_t));
@@ -51,6 +51,7 @@ int push(word_t data) {
         frame->stack_data[frame->stack_size] = data;
         return 1;
     }
+
     return -1;
 }
 
@@ -64,7 +65,7 @@ int pop() {
 }
 
 word_t tos() {
-    return (int8_t) frame->stack_data[frame->stack_size];
+    return frame->stack_data[frame->stack_size];
 }
 
 word_t *get_stack() {
